Layer Improvement... Moving things to Application Layer. Should be something like: API > depends on > Application > depends on > Domain > depends on > Infrastructure
