+++ src/Contact.Api/Contacts.Api.csproj
@@ -21,12 +21,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Controllers\" />
-    <Folder Include="Extensions\" />
-    <Folder Include="Middlewares\" />
-  </ItemGroup>
-
-  <ItemGroup>
+    <ProjectReference Include="..\Contact.Application\Contacts.Application.csproj" />
+    <ProjectReference Include="..\Contact.Domain\Contacts.Domain.csproj" />
     <ProjectReference Include="..\Contact.Infrastructure\Contacts.Infrastructure.csproj" />
   </ItemGroup>
