+++ src/Contact.Application/Contacts/Repositories/ContactRepository.cs
@@ -1,8 +1,10 @@
-﻿using Contacts.Domain.Contracts.Repositories;
-using Contacts.Domain.Entities;
+﻿using Contacts.Application.Contexts;
+using Contacts.Domain.Contacts.Models;
+using Contacts.Domain.Contacts.Repositories;
+using Contacts.Infrastructure.Repositories;
 using Microsoft.EntityFrameworkCore;
 
-namespace Contacts.Infrastructure.Repositories;
+namespace Contacts.Application.Contacts.Repositories;
 public class ContactRepository(AppDbContext context) : RepositoryBase<Contact>(context), IContactRepository
 {
     public async Task<IEnumerable<Contact>> GetByDDDAsync(string ddd)
