+++ src/Contact.Domain/Contacts.Domain.csproj
@@ -7,8 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Contracts\Services\" />
-    <Folder Include="Services\" />
+    <ProjectReference Include="..\Contact.Infrastructure\Contacts.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
