+++ src/Contact.Infrastructure/Contacts.Infrastructure.csproj
@@ -19,8 +19,4 @@
     </PackageReference>
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Contact.Domain\Contacts.Domain.csproj" />
-  </ItemGroup>
-
 </Project>
